Refactor derived metrics out of TrainSidebar
Moved wbal and VAM to a new RealtimeDataSession class
derived from RealtimeData with updateDerived method
inteded to do the computation before telemetry update
is generated and newLap method for lap data reset.
VAM is computed in metres per hour by definition,
like the VAM metric in src/Metrics/BasicRideMetrics.cpp
