+++ src/Train/RealtimeData.cpp
@@ -1095,3 +1095,42 @@ double RealtimeData::getTemp() const { return temp; }
 double RealtimeData::getCoreTemp() const { return coreTemp; }
 double RealtimeData::getSkinTemp() const { return skinTemp; }
 double RealtimeData::getHeatStrain() const { return heatStrain; }
+
+//
+// RealtimeDataSession
+//
+RealtimeDataSession::RealtimeDataSession(double CP, double WPRIME, double TAU) : CP(CP), WPRIME(WPRIME), TAU(TAU)
+{
+    // Initialize derived series data for the session
+    wbalr = 0;
+    wbal = WPRIME;
+}
+
+void RealtimeDataSession::newLap()
+{
+    // Initialize derived series data for the lap
+}
+
+void RealtimeDataSession::updateDerived()
+{
+    // Update derived series data for session and lap
+
+    //
+    // W'bal on the fly using Dave Waterworth's reformulation
+    //
+
+    // any watts expended in last 200msec?
+    double joules = double(getWatts() - CP) / 5.00f;
+    if (joules < 0) joules = 0;
+
+    // running total of replenishment
+    wbalr += joules * exp((getMsecs()/1000.00f) / TAU);
+    wbal = WPRIME - (wbalr * exp((-getMsecs()/1000.00f) / TAU));
+
+    setWbal(wbal);
+
+    //
+    // VAM
+    //
+    setVAM(vaminator.Push(getAltitude(), getMsecs(), getRouteDistance()));
+}
